requirements.txt
Django==4.0.1
djangorestframework==3.13.1
Pillow==9.0.0
